fixed CI & remove unused includes
