+++ package.repo
@@ -1,0 +1,10 @@
+repositories:
+  turtlebot3_applications:
+    type: git
+    url: https://github.com/ROBOTIS-GIT/turtlebot3_applications.git
+    version: feature-humble-panorama
+
+  turtlebot3_applications_msgs:
+    type: git
+    url: https://github.com/ROBOTIS-GIT/turtlebot3_applications_msgs.git
+    version: feature-humble-panorama
