--- turtlebot3_panorama/src/main.cpp
@@ -1,72 +1,0 @@
-// Copyright 2023 ROBOTIS CO., LTD.
-// Licensed under the BSD 3-Clause License
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//
-//   * Redistributions of source code must retain the above copyright
-//     notice, this list of conditions and the following disclaimer.
-//
-//   * Redistributions in binary form must reproduce the above copyright
-//     notice, this list of conditions and the following disclaimer in the
-//     documentation and/or other materials provided with the distribution.
-//
-//   * Neither the name of the Yujin Robot nor the names of its
-//     contributors may be used to endorse or promote products derived from
-//     this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
-
-/**
- * @file main.cpp
- * @brief Entry point for the turtlebot3 panorama application.
- * @author Gilbert, YeonSoo Noh
- */
-
-#include <memory>
-#include <string>
-
-#include "rclcpp/rclcpp.hpp"
-
-#include "turtlebot3_panorama/panorama.hpp"
-
-
-int main(int argc, char * argv[])
-{
-  setvbuf(stdout, NULL, _IONBF, BUFSIZ);
-
-  rclcpp::init(argc, argv);
-
-  rclcpp::executors::SingleThreadedExecutor executor;
-
-  std::shared_ptr<turtlebot3_panorama::PanoApp> panorama_app;
-
-  rcl_allocator_t allocator = rcl_get_default_allocator();
-  auto options =
-    rclcpp::NodeOptions(allocator)
-    .start_parameter_services(false)
-    .start_parameter_event_publisher(false);
-
-  panorama_app = std::make_shared<turtlebot3_panorama::PanoApp>(options);
-  panorama_app->setup();
-
-  executor.add_node(panorama_app);
-  executor.spin();
-
-  if (rclcpp::ok()) {
-    rclcpp::shutdown();
-  }
-
-  return 0;
-}
